+++ src/main/java/com/treasure/hunt/strategy/searcher/impl/UserControlledAngleHintSearcher.java
@@ -3,7 +3,7 @@ package com.treasure.hunt.strategy.searcher.impl;
 import com.treasure.hunt.game.mods.hideandseek.HideAndSeekSearcher;
 import com.treasure.hunt.strategy.hint.Hint;
 import com.treasure.hunt.strategy.searcher.Movement;
-import com.treasure.hunt.utils.JTSUtils;
+import com.treasure.hunt.utils.SwingUtils;
 import org.locationtech.jts.geom.Point;
 
 public class UserControlledAngleHintSearcher implements HideAndSeekSearcher<Hint> {
@@ -16,7 +16,7 @@ public class UserControlledAngleHintSearcher implements HideAndSeekSearcher<Hint
 
     @Override
     public Movement move() {
-        Point moveTo = JTSUtils.promptForPoint("Please provide a initial move location", "");
+        Point moveTo = SwingUtils.promptForPoint("Please provide a initial move location", "");
         currentPosition = moveTo;
         Movement movement = new Movement(currentPosition);
         movement.addWayPoint(moveTo);
@@ -25,7 +25,7 @@ public class UserControlledAngleHintSearcher implements HideAndSeekSearcher<Hint
 
     @Override
     public Movement move(Hint hint) {
-        Point moveTo = JTSUtils.promptForPoint("Please provide a move location", "Hint is: " + hint);
+        Point moveTo = SwingUtils.promptForPoint("Please provide a move location", "Hint is: " + hint);
         currentPosition = moveTo;
         Movement movement = new Movement(currentPosition);
         movement.addWayPoint(moveTo);
